# openapi/certifications-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Certifications.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/contracts/{contractId}/certifications": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "summary": "List certification history for an owned or administratively accessible contract",
        "operationId": "ListContractCertifications",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificationDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Certifications"
        ],
        "summary": "Create a certification for an active contract",
        "operationId": "CreateCertification",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCertificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/certifications/{certificationId}": {
      "patch": {
        "tags": [
          "Certifications"
        ],
        "summary": "Update an unfinished certification",
        "operationId": "UpdateCertification",
        "parameters": [
          {
            "name": "certificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCertificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/certifications/{certificationId}/return": {
      "post": {
        "tags": [
          "Certifications"
        ],
        "summary": "Atomically complete a certification and extend its contract",
        "operationId": "ReturnCertification",
        "parameters": [
          {
            "name": "certificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnCertificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnCertificationResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/certifications/overview": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "summary": "Get the server-filtered certification overview",
        "operationId": "GetCertificationOverview",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CertificationStatus"
            }
          },
          {
            "name": "validToFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "validToTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "name"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationOverviewRowDtoPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CertificationDto": {
        "type": "object",
        "properties": {
          "certificationId": {
            "type": "integer",
            "format": "int64"
          },
          "contractId": {
            "type": "integer",
            "format": "int64"
          },
          "assessor": {
            "type": "string",
            "nullable": true
          },
          "certificationDate": {
            "type": "string",
            "format": "date"
          },
          "protocolDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationSend": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationReturned": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CertificationOverviewRowDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "personalId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "contractId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "contractDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "effectiveValidTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/CertificationStatus"
          },
          "latestCertification": {
            "$ref": "#/components/schemas/CertificationDto"
          }
        },
        "additionalProperties": false
      },
      "CertificationOverviewRowDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificationOverviewRowDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CertificationStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ContractDto": {
        "type": "object",
        "properties": {
          "contractId": {
            "type": "integer",
            "format": "int64"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "contractDate": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "effectiveValidTo": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          },
          "prolongationWarningMonths": {
            "type": "integer",
            "format": "int32"
          },
          "prolongationAlertMonths": {
            "type": "integer",
            "format": "int32"
          },
          "prolongationForYears": {
            "type": "integer",
            "format": "int32"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/CertificationStatus"
          }
        },
        "additionalProperties": false
      },
      "CreateCertificationRequest": {
        "type": "object",
        "properties": {
          "assessor": {
            "type": "string",
            "nullable": true
          },
          "certificationDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ReturnCertificationRequest": {
        "type": "object",
        "properties": {
          "prolongationReturned": {
            "type": "string",
            "format": "date"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ReturnCertificationResultDto": {
        "type": "object",
        "properties": {
          "certification": {
            "$ref": "#/components/schemas/CertificationDto"
          },
          "contract": {
            "$ref": "#/components/schemas/ContractDto"
          }
        },
        "additionalProperties": false
      },
      "UpdateCertificationRequest": {
        "type": "object",
        "properties": {
          "assessor": {
            "type": "string",
            "nullable": true
          },
          "certificationDate": {
            "type": "string",
            "format": "date"
          },
          "protocolDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationSend": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Additional API-channel key required by every /api/v1 operation.",
        "name": "X-API-Key",
        "in": "header"
      },
      "CookieAuth": {
        "type": "apiKey",
        "description": "Secure HttpOnly authentication cookie issued by Login.",
        "name": "Certifications.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}